requirements.txt
aiohttp>=3.11.16
gradio~=5.29.0
playwright>=1.51.0
python-dotenv>=0.21.0
aiohttp-socks>=0.10.1
pocketflow>=0.0.1
pyyaml>=6.0.2
requests>=2.32.3
uvicorn>=0.34.2
fastapi>=0.115.12
duckduckgo-search>=8.0.1
colorlog>=6.9.0
beautifulsoup4>=4.13.4
pillow~=10.4.0
pydantic~=2.11.3
imagehash~=4.3.1
google-search-results>=2.4.2
chardet~=4.0.0
backgroundremover~=0.3.4
paramiko~=3.5.0
scp~=0.14.1
moviepy>=2.1.2
opencv-python~=4.9.0.80
schedule>=1.2.2
markdown2>=2.5.3
newspaper4k>=0.9.3.1
pandas~=2.1.3
ffmpeg-python~=0.2.0
lxml_html_clean>=0.4.2
nltk>=3.9.1
notion-client~=2.3.0
httpx~=0.28.1
pyimgur==0.8.1